+++ ui-next/src/test/kotlin/com/orbin/uinext/ProposalScreenshotTest.kt
@@ -303,9 +303,6 @@ class ProposalScreenshotTest {
                 onOpenFeed = {},
                 onOpenBoards = {},
                 onOpenMedia = {},
-                onOpenSearch = {},
-                onOpenDownloads = {},
-                onOpenCommands = {},
             )
         }
 
@@ -317,9 +314,6 @@ class ProposalScreenshotTest {
                 onOpenFeed = {},
                 onOpenBoards = {},
                 onOpenMedia = {},
-                onOpenSearch = {},
-                onOpenDownloads = {},
-                onOpenCommands = {},
             )
         }
 
@@ -336,9 +330,6 @@ class ProposalScreenshotTest {
                 onOpenFeed = {},
                 onOpenBoards = {},
                 onOpenMedia = {},
-                onOpenSearch = {},
-                onOpenDownloads = {},
-                onOpenCommands = {},
             )
         }
 
@@ -351,9 +342,6 @@ class ProposalScreenshotTest {
                 onOpenFeed = {},
                 onOpenBoards = {},
                 onOpenMedia = {},
-                onOpenSearch = {},
-                onOpenDownloads = {},
-                onOpenCommands = {},
             )
         }
 
@@ -367,9 +355,6 @@ class ProposalScreenshotTest {
                 onOpenFeed = {},
                 onOpenBoards = {},
                 onOpenMedia = {},
-                onOpenSearch = {},
-                onOpenDownloads = {},
-                onOpenCommands = {},
             )
         }